Move getIndicatorTableData to dataTableUtilities
